Refactor: Put mod `unix` & `windows` into separate files
